+++ packages/terra-graphs-docs/src/example-datasets/graphConfigObjects/General/timeseriesConfigXCalibrationEnabledTicksRotated.js
@@ -1,0 +1,33 @@
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+
+const getConfig = (id) => ({
+  bindTo: id,
+  axis: {
+    x: {
+      type: Carbon.helpers.AXIS_TYPE.TIME_SERIES,
+      label: 'Datetime',
+      lowerLimit: new Date(2016, 0, 1, 9, 0).toISOString(),
+      upperLimit: new Date(2016, 0, 1, 15, 59).toISOString(),
+      allowCalibration: true,
+      ticks: { tickLabelsRotation: -45 },
+    },
+    y: {
+      label: 'Line Set A',
+      lowerLimit: 10,
+      upperLimit: 30,
+    },
+    y2: {
+      show: false,
+      label: 'Line Set B',
+      lowerLimit: 0,
+      upperLimit: 250,
+    },
+  },
+  showLabel: true,
+  showLegend: true,
+  showShapes: true,
+  showVGrid: true,
+  showHGrid: true,
+});
+
+export default getConfig;